refactor: move try_init_each to token module
